accamargo-poc-laudos/requirements.txt
jupyterlab==3.0.9
matplotlib==3.3.4
pandas==1.1.5
pytest==6.2.2
seaborn==0.11.1
glob2==0.7
SQLAlchemy==1.3.24
psycopg2-binary==2.8.6
lightgbm==3.2.0
statsmodels==0.12.2
tabulate==0.8.9
plotly==4.14.3
shap==0.39.0
optuna==2.6.0
dython==0.6.4
minepy==1.2.5
pyarrow==3.0.0
boto3==1.17.41
kmodes==0.11.0
dash==1.19.0
dash-daq==0.5.0
great-expectations==0.13.17
awswrangler
nltk